overlay/base: detect bcache/mdadm/lvm storage layers
This is meant to replace the mdadm/lvm/bcache functionality of
fischstaebchen's init prog, which is mostly unimplemented anyway.

When things have been ironed out, it may be moved back to the
init prog, but for experimenting and adding new features
(luks, zram-based bcache?), a shell script is easier to write/maintain.